TaskList.java: fix findTask method to be case-insensitive
The method is case-sensitive.

A case-insensitive find is more user-friendly because users cannot be
expected to remember the exact case of the keywords.

Converting both the retrieval of task of the list and the input from the
users to lower letter case, checking and comparing of these strings
are more consistent and made case-insensitive.

Let's,
* update the search algorithm to use case-insensitive matching
* add a junit test to the new format